pyproject.toml
[project]
name = "configuration-migration-tool"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flake8>=7.2.0",
    "freezegun>=1.5.1",
    "kbcstorage>=0.9.2",
    "keboola-component>=1.6.10",
    "keboola-http-client>=1.0.1",
    "keboola-utils>=1.1.0",
    "mock>=5.2.0",
    "pydantic>=2.11.3",
    "ruff>=0.11.5",
]
